#5 feat: SortingComment CP 생성 및 CommentItem CP 단위를 확장하여 수정
