+++ src/pages/Detail/components/CommentItem.tsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 
 import ToggleButton from './ToggleButton';
+import BubbleBox from './BubbleBox';
 
 const CommentItem = () => {
   const [commentIsOpened, setCommentIsOpened] = useState(false);
@@ -9,34 +10,67 @@ const CommentItem = () => {
     setCommentIsOpened((prev) => !prev);
   };
   return (
-    <CommentItemWrapper>
-      <li>
-        <UserImage />
-      </li>
-      <CommentInfo>
-        <span>노래하는 라이언</span>
-        <span>22.10.22 16:23</span>
-      </CommentInfo>
-      <CommentText>
-        대통령은 전시·사변 또는 이에 준하는 국가비상사태에 있어서 병력으로써 군사상의 필요에 응하거나 공공의 안녕질서를
-        유지할 필요가 있을 때에는 법률이 정하는 바에 의하여 계엄을 선포할 수 있다.정당의 설립은 자유이며, 복수정당제는
-        보장된다. 정부는 예산에 변경을 가할 필요가 있을 때에는 추가경정예산안을 편성하여 국회에 제출할 수 있다. 대통령의
-        국법상 행위는 문서로써 하며, 이 문서에는 국무총리와 관계 국무위원이 부서한다. 군사에 관한 것도 또한 같다.
-      </CommentText>
-      <ToggleButton className="toggleButton" onClick={commentOpenHandler} isOpened={commentIsOpened} />
-    </CommentItemWrapper>
+    <CommnetItemWrapper>
+      {
+        <BubbleWrapper>
+          <BubbleIcon className="commentPageBubble" text="122" />
+        </BubbleWrapper>
+      }
+
+      <CommentTextWrapper>
+        <li>
+          <UserImage />
+        </li>
+        <CommentInfo>
+          <span>노래하는 라이언</span>
+          <span>22.10.22 16:23</span>
+        </CommentInfo>
+        <CommentText>
+          대통령은 전시·사변 또는 이에 준하는 국가비상사태에 있어서 병력으로써 군사상의 필요에 응하거나 공공의
+          안녕질서를 유지할 필요가 있을 때에는 법률이 정하는 바에 의하여 계엄을 선포할 수 있다.정당의 설립은 자유이며,
+          복수정당제는 보장된다. 정부는 예산에 변경을 가할 필요가 있을 때에는 추가경정예산안을 편성하여 국회에 제출할 수
+          있다. 대통령의 국법상 행위는 문서로써 하며, 이 문서에는 국무총리와 관계 국무위원이 부서한다. 군사에 관한 것도
+          또한 같다.
+        </CommentText>
+        <ToggleButton className="toggleButton" onClick={commentOpenHandler} isOpened={commentIsOpened} />
+      </CommentTextWrapper>
+    </CommnetItemWrapper>
   );
 };
 
 export default CommentItem;
 
-const CommentItemWrapper = styled.div`
+const CommnetItemWrapper = styled.li`
+  display: flex;
+  list-style: none;
+  align-items: center;
+`;
+
+const CommentTextWrapper = styled.div`
   display: flex;
   align-items: center;
   width: 100%;
+
   padding: 20px 0;
   li {
     list-style: none;
+  }
+`;
+
+const BubbleWrapper = styled.div`
+  min-width: 50px;
+  margin-right: 15px;
+`;
+const BubbleIcon = styled(BubbleBox)`
+  &::before {
+    content: '';
+    position: absolute;
+    border-top: 8px solid transparent;
+    border-left: 12px solid ${(props) => props.theme.colors.secondary2};
+    border-right: 0px solid transparent;
+    border-bottom: 5px solid transparent;
+    top: 8px;
+    left: 95%;
   }
 `;
