+++ src/pages/Detail/index.tsx
@@ -2,26 +2,20 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 
 import BookInfo from './components/BookInfo';
-import SlideButton from './components/SlideButton';
 import Comment from './components/Comment';
-import InputComment from './components/InputComment';
 import InputCommentWithPage from './components/InputCommentWithPage';
-import SlideRange from './components/SlideRange';
+import SortingComment from './components/SortingComment';
 
 const Detail = () => {
   const onClick = () => {
     console.log('this is comment test');
   };
-  const [sortIsLatest, setSortIsLatest] = useState(true);
 
   return (
     <MainWrapper>
       <ContentContainer>
         <BookInfo />
-        <SortingCommentContainer>
-          <SlideButton state={sortIsLatest} setState={setSortIsLatest} />
-          {sortIsLatest ? null : <SlideRange />}
-        </SortingCommentContainer>
+        <SortingComment />
         <Comment />
         {/* <InputComment className="대댓글" onClick={onClick} placeholder="대댓글을 입력하세요" /> */}
         <InputCommentWithPage className="댓글" onClick={onClick} placeholder="댓글을 입력하세요" />
@@ -73,20 +67,4 @@ const ContentContainer = styled.div`
   margin: 0 auto;
   padding: 24px;
   box-sizing: content-box;
-`;
-
-const SortingCommentContainer = styled.div`
-  margin-top: 40px;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const CommentConatiner = styled.div`
-  position: fixed;
-  bottom: 0;
-  width: 1024px;
-  background-color: ${(props) => props.theme.colors.white};
-  padding: 17px 26px 24px;
-  z-index: 100;
-  border: 1px solid;
 `;
